Move ComponentInfo class to a separate cs file.
