+++ src/Storage/TypeHandler.cs
@@ -10,22 +10,6 @@ using SenseNet.Tools;
 
 namespace SenseNet.ContentRepository.Storage
 {
-    public class ComponentInfo
-    {
-        public string ComponentId { get; set; }
-        public Version Version { get; set; }
-        public Version AcceptableVersion { get; set; }
-        public string Description { get; set; }
-
-        public static readonly ComponentInfo Empty = new ComponentInfo
-        {
-            ComponentId = string.Empty,
-            Version = new Version(0, 0),
-            AcceptableVersion = null,
-            Description = string.Empty
-        };
-    }
-
     public class AssemblyInfo
     {
         public string Name { get; set; }
